requirements.txt
customtkinter>=5.2.0
google-generativeai>=0.3.0
openai>=1.0.0
requests>=2.31.0
pyinstaller>=6.0.0
duckduckgo-search>=5.0.0
Pillow>=10.0.0
matplotlib>=3.8.0
